+++ js/results.js
@@ -1,5 +1,31 @@
-function loadParticipantCloud(p,experimentid,password){
-    payload={"download":p,"password":password};
+$( document ).ready(function() {
+    $('body').on('click', '.results .delete',function(e) {
+        id=$(this).parent().attr('name');
+        experimentid=$(".results").attr("id");
+        payload={"delete":id};
+        elem=$(this);
+        elem.parent().parent().hide();
+        $.ajax({
+            type: "POST",
+            data: payload,
+            //dataType: "json",
+            url: "/php/viewResults.php?experiment-id="+experimentid,
+            success: function(data){
+                //console.log(data);
+                //console.log("SUCCESS",elem);
+                elem.parent().parent().remove();
+            },
+            error: function(data){
+                //console.log("ERROR",data);
+                elem.parent().parent().show();
+            }
+        })
+    });
+});
+
+function loadParticipantCloud(p){
+    payload={"download":p};
+    experimentid=$(".results").attr("id");
     $.ajax({
         type: "POST",
         url: "/php/viewResults.php?experiment-id="+experimentid,
@@ -12,10 +38,13 @@ function loadParticipantCloud(p,experimentid,password){
         },
         success: function(data){            
             num=$(".results li").length-$(".results li .loading").length;
-            progress=0;
             name=p.split('-').slice(0, -1).join('-');
-            elem="<div class='name'><i class='fas fa-user'></i><br> #"+(num+1)+"</div>"+name+" <div class='date'>"+parseInt(100*data.progress)+"%</div>";
+            elem="<div class='name' name='"+p+"'><div class='delete' title='Delete'><i class='fas fa-trash-alt'></i></div><i class='fas fa-user'></i><br> #"+(num+1)+"</div>"+name+" <div class='date'>"+parseInt(100*data.progress)+"%</div><div class='date'>"+data.mostRecent+"</div>";
             $(".results #loading-"+num).replaceWith(elem);
+            
+            if($(".results li .loading").length==0){
+                $("#download-form").show();
+            }
         },
         error: function(data){
             console.log("ERROR");