+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [1.9.1] - 2020-07-27
+- Added option to delete participants
+- Added some error handling when loading participant data and trials
+- Fix bug with downloading results
+
 ## [1.9.0] - 2020-07-23
 - Results backend has been completely rewritten
 - Added more security measures